+++ packages/orq-rc/src/orq_ai_sdk/evals.py
@@ -3,11 +3,6 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import (
-    createevalop as models_createevalop,
-    invokeevalop as models_invokeevalop,
-    updateevalop as models_updateevalop,
-)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -213,10 +208,7 @@ class Evals(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createevalop.CreateEvalRequestBody,
-                models_createevalop.CreateEvalRequestBodyTypedDict,
-            ]
+            Union[models.CreateEvalRequestBody, models.CreateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -312,10 +304,7 @@ class Evals(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createevalop.CreateEvalRequestBody,
-                models_createevalop.CreateEvalRequestBodyTypedDict,
-            ]
+            Union[models.CreateEvalRequestBody, models.CreateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -412,10 +401,7 @@ class Evals(BaseSDK):
         *,
         id: str,
         request_body: Optional[
-            Union[
-                models_updateevalop.UpdateEvalRequestBody,
-                models_updateevalop.UpdateEvalRequestBodyTypedDict,
-            ]
+            Union[models.UpdateEvalRequestBody, models.UpdateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -520,10 +506,7 @@ class Evals(BaseSDK):
         *,
         id: str,
         request_body: Optional[
-            Union[
-                models_updateevalop.UpdateEvalRequestBody,
-                models_updateevalop.UpdateEvalRequestBodyTypedDict,
-            ]
+            Union[models.UpdateEvalRequestBody, models.UpdateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -815,8 +798,8 @@ class Evals(BaseSDK):
         retrievals: Optional[List[str]] = None,
         messages: Optional[
             Union[
-                List[models_invokeevalop.InvokeEvalMessages],
-                List[models_invokeevalop.InvokeEvalMessagesTypedDict],
+                List[models.InvokeEvalMessages],
+                List[models.InvokeEvalMessagesTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -942,8 +925,8 @@ class Evals(BaseSDK):
         retrievals: Optional[List[str]] = None,
         messages: Optional[
             Union[
-                List[models_invokeevalop.InvokeEvalMessages],
-                List[models_invokeevalop.InvokeEvalMessagesTypedDict],
+                List[models.InvokeEvalMessages],
+                List[models.InvokeEvalMessagesTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,