+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -70,7 +70,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdssbxskzs01s0zdzh1g58z2"
+        "conv_01kdt3br19pk3r2bhs03hwrqs4"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """