+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-31T08:45:55.264Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-31T11:40:33.990Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")