+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDSSBY6V7FMQS4VQWJ2TZHKX"
+        "01KDT3BREG3CGPAN900T7C3F5F"
     )
     r"""The unique identifier of the data source"""