+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -152,7 +152,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdssbxspj7ed8ywzdjzr8ah3"
+        "conv_01kdt3br1db2s7w0dn2m6atxfq"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """