+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDSSBY6R0KX81E5P08M68RFH"
+        "01KDT3BRECX8WX9V7281J2E6F8"
     )
     r"""The unique identifier of the data source"""