+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveconversationop.py
@@ -112,7 +112,7 @@ class RetrieveConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdssbxsqv1wamh9jn4ydwc0g"
+        "conv_01kdt3br1j14wxxzm359dmcs15"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """