+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDSSBY6TDMJP8NGHX2VACA6K"
+        "01KDT3BREEVPZ973AEV6TFHMMZ"
     )
     r"""The unique identifier of the data source"""