+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -150,7 +150,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdssbxssxfx6j1dt951myk7b"
+        "conv_01kdt3br1pm2355szjgxg5z4jc"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """