+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDSSBY6YN1F8FNBPG4CXC668"
+        "01KDT3BREK3NYZP9PNACH52B3T"
     )
     r"""The unique identifier of the data source"""