+++ packages/orq-rc/RELEASES.md
@@ -4469,3 +4469,13 @@ Based on:
 - [python v4.2.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.2 - packages/orq-rc
+
+## 2025-12-31 11:43:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.11 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.3 - packages/orq-rc